deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
Pillow
matplotlib
seaborn
